requirements.txt
streamlit
torch
transformers
gymnasium
scikit-learn
openai==0.28
python-dotenv
sentencepiece
